requirements.txt
google-generativeai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
rich
python-dotenv
pytest